Copy changes from gems2 to standalone ( TSolMod class, and other models) Added new files to projects s_fgl1.cpp, s_fgl3.cpp and s_fgl4.cpp Change all *.pro files and all Makefiles for Linux-GNU, Linux-PGI and CRAYXT3-catamount SD
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSTANDALONE/trunk@288 9ff7d5ee-7a19-0410-ab54-ec97d5369f81

+++ GEMIPM2K/main.cpp
@@ -33,7 +33,7 @@
 int main( int argc, char* argv[] )
  {
    double PureTime = 0.0;
-   int nNodes = 1, nRecipes = 0;
+   long int nNodes = 1, nRecipes = 0;
    char  (*recipes)[fileNameLength] = 0;
 
    //  char *recipes = NULL;
@@ -66,10 +66,10 @@ int main( int argc, char* argv[] )
 
   // Number of ICs, DCs, Phases and Phases-solutions kept in the node
   // DATABR structure for exchange with GEMIPM - for your convenience
-  int nIC, nDC, nPH, nPS;
-  int i,   j,   k,   ks;    // indices for direct access to components
+  long int nIC, nDC, nPH, nPS;
+  long int i,   j,   k,   ks;    // indices for direct access to components
                             // and phases data in the DataCH framework
-  short nodeHandle, NodeStatusCH, IterDone;
+  long int nodeHandle, NodeStatusCH, IterDone;
   double NewMass;
   
   // Getting direct access to DataCH structure in GEMIPM2K memory
@@ -137,7 +137,7 @@ int main( int argc, char* argv[] )
     	 goto FINISH;
      
  	  
- 	 for(int cRecipe=0; cRecipe < nRecipes; cRecipe++ )
+ 	 for(long int cRecipe=0; cRecipe < nRecipes; cRecipe++ )
      { 
         // Trying to read the next file name 
  	    sprintf(NextRecipeFileName , "%s%s", input_recipes_file_list_path, recipes[cRecipe] );
